requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.0
python-multipart>=0.0.5
aiofiles>=0.7.0
playwright>=1.40.0
beautifulsoup4>=4.9.3
httpx>=0.18.0
python-jose>=3.3.0
passlib>=1.7.4
pydantic>=1.8.2
email-validator>=1.1.3
python-dateutil>=2.8.2
pillow>=8.3.1

# Additional playwright dependencies
pytest-playwright>=0.4.0
greenlet>=2.0.0  # Required for async support